Proper pol interleaving in dsim
Both polarisations now share a stream, with a higher-level class
wrapping a set of polarisations. This is enabled by the (not-yet-released)
spead2 feature of specifying an endpoint per packet.

It also now supports specifying multiple addresses per polarisation.

The infrastructure is in place to support multiple digitisers
(each in their own thread, with loose synchronisation) but it's not
exposed.