Fix max packet size in receiver
It was using the chunk size instead of the packet size. When using
multi-packet receive queues it doesn't actually matter, because the size
is just used to reject packets that are too large. But if MPRQs aren't
available for whatever reason, there would be far less space in the
buffer.