Add --use-gdrcopy option
When used, it uses [gdrcopy](https://github.com/NVIDIA/gdrcopy) to
unpack packets directly into GPU memory, which eliminates a read and a
write from CPU memory. It doesn't seem to have much effect on CPU load
for one antenna, but should allow more antennas to be packed in.

It depends on a gdrcopy Python wrapper package, which I'm still
developing (but it is only needed when actually passing the option).