Initial PeerDirect support
It's a bit clunky because the queuing mechanisms were never designed for
it. To make it minimally intrusive, there are still separate types to
hold spectra being computed versus transmitted, and the copy step is
replaced by a swap.

It runs, but I haven't examined the output for sanity yet.