Improve documentation for katfgpu
In this commit I have documented just about everything. Probably not
perfectly, there is likely still room for improvement, but I hope that
the barrier to entry for new developers is now lower.

I've added Sphinx documentation to what little there was, so now each
class and function shows up in a "Reference Manual" of sorts.

No functional changes to the code are applied in this commit, only
documentation. And of course `black` had its way with the quotation
marks, line length, indentation, and what have you.

Resolves: NGC-175