Add dsim compilation to Docker build (#18)
* Add dsim compilation to Docker build

The dsim depends on spead2 so that needed to be built and installed
separately from katfgpu itself on the build image. I'm not sure if
there's a more efficient way to do this.

I guess it's not strictly necessary but I also brought the version of
spead2 included more up-to-date (v3.2.2). The old spead2 commit
that was referenced pulled in and built a different (older) version of
numpy each time the build was run, which almost doubled the
time that it took to build. It doesn't do this now.