Implement fringe phase into delay model and postproc (PR#17)
Mathematically this is very similar to the delay and delay rate on the
host side. In the GPU, it's just a constant factor that gets added to
the fine-delay across the entire spectrum. It may not be implemented
correctly at the moment, as it should be done from the centre of the
band and this commit does it from the bottom. Add a TODO note to the
postproc's docstrings to remind me to fix this.


AFAIK this brings the GPU F-engine more or less to feature-parity with
the SKARAB one. One issue that I'm not confident about is that the fine
delay and fringe phase are both calculated with the earliest timestamp
of the spectrum, where I'd expect it to be preferable to use the middle
one. But according to Andrew Martens, this is how MeerKAT currently does
things.

Also update the the postproc unit-test to incorporate fringe phase.
May the pre-commit gods forgive me for not PEP8-ifying the rest of
the unit test file.

Update comments claryfying how the delay model range inversion works.
I hope that this update helps a bit. I wrote the comments in question
originally but I had to scratch my head for a bit to figure out what
was going on.

Contributes to: NGC-181

Co-authored-by: Bruce Merry <1963944+bmerry@users.noreply.github.com>