Add basic versioning capability
I've never tried using setuptools_scm, but following the directions on
the web seemed straightforward enough.

Mypy messes things up somewhat, so I had to put `type: ignore` things in
a few places.

Also add a --version argument to the command line, and a
requirements.txt in the doc/ folder to make building the docs a bit
easier.