Improve aiokatcp implementation
Move aiokatcp to a more aesthetically pleasing location on the
requirements.txt.

Add katcp-port as command-line parameter. Set the default to 7147. I am
not sure what the convention is but if you're just testing an F-engine
alone, you don't need to specify something which seems sane. Also add it
to the docstring of the Engine class, and remove the default argument
there.

Change listening interface on Engine from "localhost" to "" to listen on
all interfaces.

Change auto strategy on dropped-packets sensor from default to
EVENT_RATE.