Further refine versioning infrastructure
The [setuptools documentation](https://setuptools.readthedocs.io/en/latest/pkg_resources.html)
discourages the use of pkg_resources directly as there's performance overhead
at runtime. Similarly, [usage in Sphinx](https://github.com/pypa/setuptools_scm#usage-from-sphinx)
is discouraged. Both references suggest `importlib.metadata` instead.

It also seems strange to me to require setuptools-scm to be installed, as
this is something that pip pulls in to the build environment when it is
needed. So I have removed requirements for it, and references to it in the
code, opting for `importlib` as recommended.

I've also removed the `setup_requires` kwarg from `setup.py`, in favour of
the list of build requirements in `pyproject.toml`.

Identifiers for editable installs were [rejected from PEP660](https://www.python.org/dev/peps/pep-0660/#editable-local-version-identifier),
at least as far as annotating the `version` specifier goes. This breaks
things with pip.

For versioning from within editable installs, some others seem to have
implemented something [vaguely similar to katversion](https://github.com/pypa/setuptools_scm/issues/273#issuecomment-519473083).

The maintainer of setuptools_scm says that [there's no good solution to
the editable installs problem right now](https://github.com/pypa/setuptools_scm/issues/518).

We could perhaps consider a (clunky) workaround like [this](https://stackoverflow.com/questions/43348746/how-to-detect-if-module-is-installed-in-editable-mode#57854737).

My personal feeling is that an editable install is an exceptional case
that we shouldn't optimise for - I think the onus is on the developer to
remember that they are working in one, and not let the ambiguity get the
better of them. But I am willing to consider other points of view on this.