Remove types.py imports (#26)
* Remove types.py imports

In favour of the sanctioned ABCs in katsdpsigproc.

However, because in the FFT operation, we refer to the _pycuda_stream
directly, mypy throws its toys. I've marked to ignore that line for now
but we need to make a decision what to do for future.

Also updated a docstring making use of the removed types import.