Add isort to pre-commit flow (#26)
A number of files had comments like "1. Import local modules" and after
isort they were ending up in odd places or the numbering was
out-of-order, so I deleted them. With isort there shouldn't be any need
for such comments because imports will always be in a standardised
order. It does mean that Gareth's numbered comments in some files now
start at 2 or 3, but that's why you shouldn't hard-code such numbering
unless there is an extremely good reason.