Unify requirements and pyproject.toml
Merge requirements.txt files from both projects into a requirements.in, and
use pip-compile to specify a completely reproducible environment. This seemed
like a good development practice. Also add a little script to help a new cloner
get set up to develop quickly.

Create unified pyproject.toml. It was more or less the same as that of katfgpu.

Merge gitignore file between the two former repos. Add a venv to the list. Use
`sort` to get it more or less in order. Punctuation makes it weird though.