Add a 'test' extras_require
This serves as executable documentation of the packages *needed* to run
the unit tests (executable as `pip install .[test]`), as opposed to
requirements-dev.in which contains other packages like pre-commit that
are not directly related to the unit tests.

I'm assuming that pytest-xdist is listed in requirements-dev.in just
because it can help speed things up, rather than because it is a
requirement.

Relates to NGC-243.