Make mypy pre-commit step pass
It's done by just excluding the xbgpu code, which will need to be worked
on.

Also added some missing dependencies (types-six and types-decorator)
which mypy started complaining about for some reason, and added some
config options to make mypy.ini the sole source of truth about which
files to check.