Add pytest.ini to root
Remove the misc ones from the individual directories.
Adjust test parameters to remove the *really* long tests, >84 antennas,
because it'll be a long time before we really need to support that. Also
add 8k to list of channelisations because that's been talked about.

Also remove old Dockerfiles because they have already been merged.