Make xbgpu pass mypy type-checking
Remove the line from mypy.ini that excludes xbgpu's code and tests from
checking. It all passes now.

Let mypy ignore the instances where we manually alter the dtype of an array.
There doesn't seem to be a way to get around this.

When altering the value of numpy-managed signed-integer variables, explicitly
create the right type. I'm not sure whether this is really the right approach
though, so I've left a note indicating as much.

Remove verify_reorder() from the precorrelation reorder test. It's superceded
by the ported-to-C function, and we can get it back from the git history if
we need to. It doesn't seem worth the effort (to me) to get it working if it's
not getting used anyway.

Change AbstractStream to _AsyncStream to reflect what is available in spead2's
stub files more accurately.

Change some inconsistent usages of cuda.Context as a type hint where something
marked as an AbstractContext is being passed in. You can give something
more specific as an argument when the general case is type-hinted, but not
the other way round. Also while I'm there, remove redundant type hints from
the docstring.

Resolves: NGC-256