Fix xbgpu documentation in response to feedback
Move networking section of xbgpu to the root rather than as a subsection
of the theory document.

Run references.bib through
`biber --tool --validate-datamodel <filename>.bib`.
Turned out not to be the problem but it seemed reasonable to include.

Remove `references.bib` from the bibliography directive used in pfb.py.
This turned out to be the error. I don't know why it worked with katfgpu.
Replace with a list of actual citations that we want to list. Cleaner.

Include proper section links in the body of the xbgpu documentation,
which was once markdown.

Replace Monitor references in type hints so that proper references can
be made. I know it's going to be abstracted away soon enough but I like
not having any warnings shown.

Also fix up some indentation issues in xbgpu's ringbuffer module doc.