Move dsim and fsim out of scratch directory
I omitted `-pipe` from the Makefile because I saw no evidence that it
sped things up (if it did, it was lost in the noise).

I changed the scratch/fgpu Makefile to build memcpy_loop, since that was
the last .cpp file in that directory, although possibly it belongs more
in dc_sand. The xbgpu Makefile is removed since the rest was just
Python.

Closes NGC-264.