Introduce unit-testing for fgpu's katcp interface
In this commit, I introduce unit-testing for fgpu's KATCP interface. So
far I have only added tests for the "requests" aspect, not the sensors
just yet.

Initially, I have included a test for the quant-scale, because this
functionality is already implemented. It passes. I also include a test
for delays, inspired by the CBF-CAM ICD (though this element will talk
to a product controller, not CAM directly). This test does not pass,
because the functionality hasn't been implemented yet. I'd like to get
it reviewed for sanity.

Also add some units to the docstrings in delay.py, for clarity.

Addresses: NGC-227