Don't test every combination of parameters by default
Greatly reduce the duration of the tests by changing the xbgpu tests
that parametrize on three dimensions to only run enough tests to cover
each value of each individual parameter, rather than all possible
combinations. This cuts the test duration on my machine to about 3
minutes (or half that if I run `pytest -n 3`), which is still on the
high side but manageable.

Add a `--all-combinations` argument to pytest that can be used to
restore the original behaviour, which could be used in nightly testing
for example.