Refine pytest-coverage-related things
Small udpates based on PR feedback.
I would still like to move the configuration to the pyproject.toml file, and
there's still work to do in getting the configuration as complete as
possible so that the user can just run a simple command to do the tests
and still get useful output. But that's more for tomorrow.