Update dev-guide.rst
The text incorporates the latest and greatest of what I understand of
the unit-testing process.

I also remove a redundant line from pyproject.toml, which specified the
dynamic context for coverage, but it gets overridden by the
`--cov-context` option of `pytest-cov` anyway.

I also change the reference to the `coverage` documentation from
specifically pointing to 5.5 to rather point to stable.