Add --rx-reorder-tol to control number of active chunks indirectly
Instead of directly specifying the number of active chunks, specify a
maximum variation in arrival time, in units of samples. This should
translate directly into memory usage for a given number of antennas and
X-engines, independent of the number of channels or number of batches
per chunk (at least for wideband), and it also has an intuitive
relationship to time in seconds. I didn't make it use seconds, (a) to
emphasise that it is related to the ADC timestamps in packets rather
than the passage of time, and (b) so that memory usage won't depend on
sample rate; but I could be convinced otherwise.

I'm not wild about the name of the command-line option, but it can
probably be revisited as part of NGC-226.

See NGC-229.