Streamline the requirements files in the repo
In this commit I merge the requirements-dev.in and requirements-doc.in.
It seems to me that they don't really need to be separate.

I also add `-c requirements.txt` and `-e .` to requirements-dev.in,
which means we can do just one `pip install -r ` command in
dev-setup.sh.

I also added ipython as a dev requirement because many people like it,
so it makes sense to have it around. And if it's there, `pip-sync` won't
remove it.

I think this is overall an improvement because fewer files to maintain
is an advantage.