Adjust delay-model control based on PR feedback
Change start_time in the delay request to an aiokatcp.Timestamp type,
it's a wrapper around a float.

Make --sync-epoch a required cli arg.

Rename gpu_context fixture to device_context, to be more in harmony with
katsdpsigproc's conventions. Adjust the getting of host and port from the
engine_server fixture in engine_client to accommodate for future IPv6
possibility.

Make SYNC_EPOCH "const" in test_katcp_interface. Should be less
opportunity for bugs to creep in. Also parameterise malformed delay strs,
so that the test can do a few combinations of them. Testing is pretty
thorough now.