Extend engine_server fixture for extra args
You can now mark a test with @pytest.mark.engine_arglist_override and
add extra arguments (possibly overriding previously-given ones in the
class).

It's not used as yet, but I think it may well be as we develop the unit
tests for fgpu for fgpu for fgpu for fgpu for fgpu for fgpu for fgpu for
fgpu for fgpu a bit more.