Use pip-compile in pre-commit
This ensures that requirements.txt and requirements-dev.txt are always
up to date with respect to the corresponding .in files. In particular,
if a requirement is updated in requirements.txt, it will ensure that
requirements-dev.txt gets the same version (if it appears in both).

Removed pip-tools from dev-setup.py (no longer used there) but added it
to requirements-dev.txt so that one can call it directly to update
requirements.