Change pre-correlation reorder to use numba
In this commit I remove the dependency on an externally-compiled library
for the pre-correlation reorder unit tests. I include a naive, simple
reorder using for-loops in Python then accellerate it with Numba. Based
on my testing, it seems to be about 2-2.5x faster than the previous
implementation.

Additionally, I noticed that a lot of the comments and docstrings had
become out-of-sync with what was actually being done. I removed a lot of
it, and tried to make the code as self-explanatory as possible.

Resolves: NGC-297