Port xbengine test to numba
It currently works and is about 2x as fast as Gareth's C code.
It's still ugly, may the pre-commit gods forgive me, but I wanted to
commit and push this in case something terrible happens overnight
because it represents several hours worth of work.

Still todo is removing the C code, adjusting the dev-setup script and
cleanup / docstrings.

Resolves NGC-298