Use exact dimensions in precorr reorder's slots
In this commit I change the dimensions in the
PrecorrelationReorderTemplate to accel.Dimension instead of just ints,
and I specify exact=True. This is because the actual kernel uses some
interesting pointer arithmetic and doesn't check boundaries much (as far
as I can see anyway). I feel that it's safer this way.

I also adjust extend the dimensionality by 1 and change the type to
np.int8 instead of np.int16, to represent the complex numbers. This is
the convention that we are working towards where we have to send a SPEAD
stream with this shape out, so it makes sense to be consistent from
earlier on in the pipeline.

I also modify the way that the precorrelation reorder information is
used in xbengine.py, namely that buffers created for the various
EventItems inspect the IOSlot's shape and dtype rather than getting the
shape from the template and hardcoding the dtype. This seems more robust
to me.

Mypy complains about this because it's checking against IOSlotBase,
which doesn't have shape and dtype attributes, so I have just silenced
it for these lines for the time being.

fgpu does this much better in accordance with the DRY principle, so I've
created NGC-308 in order to remind us to abstract these things.

Additionally, remove the suffix `_operation` from
`self.precorrelation_reorder`, because it makes the lines just a touch
shorter and easier to read.

Working towards NGC-225 (eventually).