Add more input sensors to fgpu
Add sensors for input bytes, heaps and chunks. The
input-missing-heaps-total sensor is updated to count any reasons for
heaps not making it to the pipeline, whether because they were missing
from a chunk, the whole chunk was missing, or the chunk's pol-partner
was missing. This does mean that if one pol completely stops
transmitting, the sensors will all stop updating as they are now only
updated when a pair (across pols) is yielded.

There are also some tweaks to avoid accessing `chunk.present`, which
would involve copying the data from a `vector<bool>` to a Python list
(and it was accessed twice!) The sum is now computed in C++ via a new
`n_present` property. I'd hoped this would reduce CPU load, but it
doesn't seem to have made any difference.

The original warning log for lost heaps is downgraded to debug, to avoid
flooding the logs when things go bad (which in turn can cause more
dropped heaps due to the logging overheads). Set KATSDP_LOG_LEVEL=DEBUG
in the environment to see them.

See NGC-285.