Change the dimensions of the tc-correlator kernel
The changes in tensorcore_xengine_core.py are relatively simple.
Highlights including making the dimensions explicitly exact, and
adjusting the dtypes to actually reflect reality, due to the addition of
the complexity dimension. I also re-flowed the docstrings to make them a
bit nicer. I clean up the documentation a bit as well.*

Following on from this, in xbengine.py I make the shapes and dtypes of
the buffers in the QueueItems get their info from the actual slots of
the operations, rather than repeating that information here again. I
also remove _operation from the names of the operation objects, just to
make things a bit more readable IMO. And I remove some of Gareth's
unnecessary assignmentless declarations and type-hints.

I adjust the output heap shape, and datatype, to match that coming out
of the correlator kernel, and it should also be inline with the ICD now.
This was the main thrust of this effort, the other changes that had to
be done were just in consequence of this.
I also adjust the data types and expected sizes in the appropriate unit
tests.

The unit tests for the correlator had some baked-in assumptions about
the data shapes and sizes. Try as I might, I just couldn't get both of
them working at the same time. The "exhaustive" one was simpler to get
working with what I consider to be a "correct" output shape, but the
"quick" one just wouldn't, unless I went back to the old
(baselines*pols*pols) shape assumptions. Which broke everything else. So
I rewrote the correlation unit test to be equivalent to the "exhaustive"
one, but also quick, because numba can do its thing in parallel. Which
is amazing.

* I also add sphinx.ext.imgmath to the list of Sphinx extensions, and I
sort the extension list into alphabetical order for good measure.

Resolves: NGC-225