Add katcp and network sensors to xbgpu
Modify XBEngine to inherit from aiokatcp.DeviceServer. Add katcp host
and port arguments to its list as well as to main's cli arguments. Remove
some of Gareth's totally redundant comments from the parse_args function.

Add network-related sensors, to bring xbgpu more or less to parity with
fgpu in terms of the network analytics that it can give to Prometheus.

I also tidy up the launching of the three "loops" in the xbengine, because
they were annoying me. And I reworked the documentation for the XBEngine
class, while I was adding parameters. I removed __init__'s docstring,
and re-flowed it to 80 characters instead of 120.

Resolves NGC-286