Add num_ants to fgpu's SPEAD sender for heap sequence
Turns out that the F-engines kind of do need to know how many antennas
there are in an array, because their output heap counters need to line
up.

The spead2 receiver on the X-engine side doesn't read into the actual
items, just the first 4 mandatory ones, for heap reassembly, so it
can't tell packets from antenna A or B if they're all labelled with a
heap_counter X. Which is highly likely if we start all the F-engines at
more or less the same time.

So I have given the spead2 senders knowledge not just of their fengine
ids (Which they had previously) but also how many F-engines there are,
so that they can give subsequent heaps appropriate increments in their
counters.

Resolves: NGC-338