Quick-n-dirty factory for chunk_place
Instead of passing layout parameters to the callback, compile a callback
that is specialised for given layout parameters. This lets LLVM do much
better optimisation e.g. use shifts instead of divisions because it
knows the denominator at compile time.

It still needs some cleanup: the layout can now become a regular Python
class instead of a numpy structure, and it's compiling the code twice
(once for each stream).