Make provision for 'Transmission Rate Factor'
As mentioned in NGC-269, there is a need to allow for adjustment of the
transmission rate of the Engines to account for any network jitter, etc.
The variable name is certainly up for discussion, this initial effort
was done to ensure the value propagated wherever necessary. There is
also an update to the run-fgpu script to reflect the updated ADC rate
variable name.

No error-checking of this value has been done just yet; it is assumed
anyone adjusting the value is aware of the consequences. This has been
tested on fgpu (using a dsim) and xbgpu (using an fsim) and the
pipelines do stall as expected (especially for low values, e.g. 0.5).

Contributes to: NGC-269.