Remove checkout stage in Jenkinsfile
After much debugging this Jenkinsfile is finally working. The checkout
stage was removed from the Jenkinsfile, as the checkout is automatically
done when the Jenkins job is configured as a multibranch pipeline. The
configuration of the Jenkins job must enable 'Recursively update submodules'
to pull in the spead2 repo. If this is not done the the Jenkins pipeline
will fail in the 'Install katxbgpu package' stage.