xbgpu: wait for events from free items
This fixes a race condition: the gpu_proc_loop was returning the rx item
to the receiver loop without waiting for the precorrelation reorder
kernel (which reads from the rx item) to complete. Fix it by

- setting an event on the rx item when returning it
- waiting for the event in the receiver loop

Also wait for any events on items retrieved from `tx_free_item_queue`
just to be safe, although at present there are none (and none are
needed).

Closes NGC-363.