Update tensor-core kernel to open-sourced version
It now matches the code at
https://git.astron.nl/RD/tensor-core-correlator/, revision
83abdccb9377510427e91068ead6a30d0124ac44.

Compared to the previous version:
- references to "station" are renamed to "receiver" (this was already
  the case in upstream v0.5, but was not part of our modified v0.5).
- most of the type-punning has been replaced by memcpy's to give defined
  behaviour.
- The main entrypoint is now directly marked `extern "C"` rather than
  wrapped in an `extern "C"` block.
- I put back the calculation of NR_BASELINES from upstream, replacing
  the option to pass it in as a parameter (since it cannot sensibly be
  set separately from the number of receivers).

I also modified the way the parameters are passed in: instead of using
mako templating, they're passed in by `-D` compiler command-line
arguments, as the upstream version intended. This further reduces the
number of changes compared to upstream. This will allow the file to be
renamed from .mako to .cu since it no longer needs any templating.

Closes NGC-374.