Rename tensor_core_correlation_kernel.mako to .cu
I removed all the mako templating, so renaming it to .cu is appropriate
(and helps editors apply suitable syntax highlighting, for example).

I also updated the way it is compiled so that it no longer runs through
the mako frontend of accel.build, calling context.compile directly; and
using the more modern importlib.resources to load the code.