WIP on NGC-254
This provides the bulk of the implementation. The existing C++ module is
removed and replaced by Python code. The chunk class pre-creates its
heaps for efficiency, using some careful coding to allow the timestamp
to be updated in place. The ringbuffers between run_transmit and the
actual sending are gone. Instead, a free queue is stored in the Process
class. There is no matching data queue, as the queue inside the spead2
Stream object is the partner.

There are still a number of things to take care of:
- Update the documentation, particularly the design docs.
- Fix up the PeerDirect support. It replies on swapping GPU buffers
  between OutItem and Chunk, which won't work any more now that Chunks
  have pre-assigned heaps. It will probably need to be re-designed from
  scratch.
- Get rid of the spead2 submodule entirely. For now it's left in place
  because the Dockerfile uses it for building dsim and fsim.
- Some proper end-to-end unit tests.