Eliminate 3rdparty directory
spead2 C++ bindings are now fetched from Github instead of the
submodule. Since releases are already pre-bootstrapped this simplifies
things as autoconf, automake, jinja2, pycparser are no longer required.

The spead2 bindings are put in the build-base target of the Dockerfile,
which removes the need for Jenkins to install them again.