Initial work on fgpu end-to-end test
This just provides some scaffolding and pushes data though the system,
but doesn't do any verification of the results. It currently relies on a
sleep to know when to stop; it will need NGC-277 to be fixed to be able
to reliably know when the data has stopped flowing.

Relates to NGC-381.