Cleaner shutdown path
This is mostly to facilitate unit testing by making it possible to push
data through the pipeline and get notification (an end control item in
the output stream) when it has all been processed.

The lifecycle management is integrated into that provided by aiokatcp.
Instead of a `run` method explicitly called from `main`, `start` is
overridden to start the processing in the background, and `on_stop` is
overridden to wait for it to complete. When a signal is received, the
input streams are stopped, and each stage then passes a `None` object to
the following one to tell it to stop.

Closes NGC-277.