Fix output chunk correction when data arrives too quickly
If a chunk didn't finish transmitting before the next chunk was taken
off the free queue, the done callback would return the wrong chunk to
the free queue due to the way closures work in Python (they capture the
variable rather than the value the variable had when the closure was
defined).

To avoid any such issues with closures, the done callback is pulled out
entirely into a new method and values that need to be captured are done
so explicitly with `functools.partial`.