Fix adjustment to phase_rate from katcp
This is an addition to bb05b338c335afc370acdc791781b97eec6d0732, which
fixed calculation of the initial phase term, to also adjust the phase
rate based on the delay rate.

Also removed the interface-only test of the `?delays` command that
checked the internals of the delay model. It would have been invalidated
by this change, and this is better tested by actually pushing data
through the pipeline (for which some tests have been written, with more
coming).

Relates to NGC-296.