Fix SPEAD item names used in xbgpu testing
It doesn't actually affect the test at all (since xbgpu works off the
item IDs), but it irked me that the xbgpu tests used non-standard names
for the SPEAD items.

I also deleted the padding the test was adding. The mechanism is was
using for the padding (inserting immediates with non-zero item IDs)
differed from the standard mechanism of using ID 0 (which is reserved
for padding) and it was easier just to remove it than try to fix it by
manually building heaps. The padding has no influence on the code in
xbgpu, being handled transparently by spead2.

Closes NGC-362.