Fix calculation of send rate for fsim
It was arbitrarily and needlessly making assumptions about the number of
multicast streams based on the number of antennas, when it can be
determined from the number of channels per stream.

I also fixed the display of the output rate to use SI units for Gbps,
which I believe is standard in networking.

Partly closes NGC-423.