Add Prometheus metrics export to fgpu
Currently there are no useful metrics - it is just proof-of-concept for
NGC-325.

The make_engine function was modified to take the pre-parsed
command-line arguments. This makes allowed async_main access to the
arguments, and has the side benefit that running with `--help` will not
try to create a device context.