Move output metrics into send module
Also increment the counters as each frame is sent, rather than once for
the whole chunk, to more accurately handle any cases where a frame fails
to be sent. I'm a bit nervous about the cost of this when there are few
channels, but we shall see as we go.