Make PFB filter weights symmetric
The sinc was half a sample offset from the Hann window. Shift it by half
a sample to line them up. This makes very little difference in practice,
but ensures that its Fourier transform is purely real (for an
appropriate choice of origin), and hence avoids odd phase responses. See
the analysis in scratch/fgpu/pfb_weights.ipynb for more details.

Closes NGC-278.