Make xbgpu report network stats via Prom counters
Some refactoring was required because the tests broke down when both
fgpu and xbgpu tried to register counters called input_heaps. To cope
with that, the METRICS_NAMESPACE is changed from the generic `kat` to be
`fgpu` or `xbgpu` (I also considered moving the input counters up a
level in the hierarchy, but they have different labels so would still
have conflicted).

Also removed the logging of dropped heaps, to match fgpu. If an f-engine
fails this can make the logs extremely noisy very quickly.

See NGC-325.