Bring back some monitoring of the spead2 ringbuffer
It's not quite as capable as the original C++ code, because there is no
callback generated when spead2 pushes a chunk into the ringbuffer. The
queue size is polled immediately after removing an item, and hence forms
a lower bound (which should be sufficient to diagnose back-pressure).

Also removed free_chunks event_qsize call - there is no tracking of the
free ringbuffer. It would be useful, but isn't very practical as there
is no way to know when items are removed (and in particular, when it
runs empty).

Relates to NGC-377.