Per-channel, per-pol complex gains
Implement equalisation gains on a per-channel and per-pol basis, instead
of having just a "quantisation" gain. Since these gains are no longer
just about quantisation, the arguments, commands, variables etc have
"quant" removed. The katcp interface is modelled on the MeerKAT CBF-CAM
ICD, but with the input replaced by a number in commands and by "input0"
or "input1" in sensors.

Internally the implementation is kept simple rather than efficient.
Every output item gets a copy of the gains and copies it to the GPU,
even if there are no changes, which avoids any need to prevent race
conditions or track dirtiness.

Closes NGC-390.