Extend test_correlator to test addition and saturation
The main goal is to test saturation (closing NGC-307), but it more
generally tests that the computed visibilities are added to the existing
values. This probably makes test_multikernel_accumulation redundant (see
NGC-329) but I've left it alone for now.

The unit tests fail because the saturation is not yet implemented.