Add misc documentation updates
As per feedback on PR #131, there were a few formatting changes required
to conform a bit better to the .rst format. The X-Engine packet format
diagram was updated with the correct byte increments. I still need to
check if the relative links to documentation is working as expected.

Contributes to: NGC-376.