Revert the asyncio.gather in xbgpu/test_engine
I had forgotten to await the gather, and adding that in resulted in the
process seemingly free-running. It has been rolled back to await'ing the
recv_process before cancelling the server.join.

@bmerry informed me of the pending NGC-441 which would make things
easier, but I should finish this first before moving on.

I will attach a screenshot to the ticket comparing pytest runs on main
versus xbgpu_shutdown_cleanup.

Contributes to: NGC-276.