dsim: refactor sending code
Move a lot of code from main.py to send.py. Also introduce the HeapSet
class to bundle the data for a heapset with the partitions, and add a
separate Sender class to manage the futures. This design should
facilitate switching out the HeapSet in a later version.

Also adds signal handlers so that Ctrl-C shuts down cleanly.