dsim: speed up CW generation
Instead of using np.cos on a linear array, use repeated doubling to
generate regularly-spaced samples (similar to the trick used in an FFT).
This means that only O(log n) transcendental function evaluations are
required, with most of the work done with just complex multiplies, which
can be done in single precision with reasonable accuracy.

For very large timestamps this may actually improve accuracy, since the
original code would have lost a lot of accuracy in the range reduction.
With the new code everything is computed relative to the first
timestamp, so range reduction errors only affect the overall phase of
the signal but won't distort it.

The result is that sampling the CW is about 5x faster, and I believe
should use less memory too.

Relates to NGC-456.