dsim: rework CW signal generation
The previous approach was clever because it reused an chunk-sized table
of complex exponentials multiple times, but the dask scheduler seemed to
get confused by it, and it used excessive amounts of memory.

Rework things so that each chunk is completely independent. This is much
easier for the scheduler to do the right thing with.

Performance should still be similar, as the asymptotic number of flops
doesn't change, and it seems to do similarly in practice.

Relates to NGC-456.