Improve synchronisation between disparate dsims
Previously the synchronisation logic came before generating the initial
signal, which can take a system-dependent amount of time and thus lose
synchronisation. Now the synchronisation is moved closer to the actual
start of transmission.

To make this viable, the signal has to be generated for a fixed initial
phase (0) rather than basing the phase on the first timestamp. That in
turn means the first timestamp has to be a multiple of the whole signal
length, rather than of a single heap. To compensate for the rounding
error involved in this alignment, a sleep is added to align real time
with the chosen first timestamp.

Closes NGC-470.