dsim: add a Prometheus gauge tracking time error
It measures the difference between the time specified by the current ADC
timestamps and wall clock time, which will make it obvious when the dsim
is not keeping up with the required rate.

Partially addresses NGC-473 (it will still need a Grafana dashboard to
show it, and possibly raise an alarm).