Split input and output chunk sizes
The single `--chunk-samples` command-line option is replaced by
`--input-chunk-samples` and `--output-chunk-jones`. The minimum output
chunk size is one frame, which for 32K channels and 256 spectra per heap
corresponds to 32MiB per chunk. To support vkgdr on GeForce GPUs without
ReBAR it's necessary for the input chunks to be much smaller than this.

The internals already supported this, because coarse delays already
ensured that input and output couldn't be in lockstep. So this is just
the plumbing around command-line argument processing.